Implementation of the new $solr_host setting.
-------

New solr->host setting. It has been added to make sure that people can use a solr instance installed on another server than the "main" structWSF instance server.

By default, in the WebService.php file, the value is to "localhost". This means that even if this file is not updated on an installed version of structWSF, it won't impact a running instance.

However, I encourage everybody to add this new setting to any running data.ini files.