A few things got enhanced:
Now, to comply with the SPARQL protocol, if no "dataset" parameter is defined, then we check if we can find a "default-graph-uri" or "named-graph-uri" parameter and to use it as the "dataset" parameter.

Also, if no "dataset" parameter is defined, we don't return a 201 error right away. We check if some graph URI has been defined within the SPARQL query. If the user did, then we check if he is has access to it. If he doesn't, or if no such dataset URI are defined in the sparql query, *then* we send the 201 error.

Also note that a CONSTRUCT query still return a 201 error since the current SPARQL query analisys tool is not yet evolved enough to check, and analyse, the "graph" clauses within the CONSTRUCT (nor SELECT) SPARQL queries. Instead, change the Content-Type mime wanted in order to get back the resultset in some serialization format.

One minor thing got fixed:

- Some XML encoding when generating some RDF/XML serialization