(1) Addition of a new parameter:
"aggregate_attributes": Specify a set of attributes URI for which we want their aggregated values. The URIs should be url-encoded. Each attribute for which we want the aggregated values should be seperated by a semi-colon ";".

This new parameter enable someones to list all the values, and their count, for a given attribute. This is used to extend the filtering capatilities of the endpoint.

(2) Fix prefixes creation from namespaces used to describe records of a resultset returned by the endpoint.

(3) Solr encode some of the values

(4) Sort results by their counts

(5) Adding a new autocompletion behavior for the prefLabel. This new autocompletion behavior is trigged by the double stars markup (**) when used as a value on a prefLabel attribute value filtering query to the endpoint. This new markup makes sure that the solr query is properly created and modified so that we can create autocompletion on multiple words, including spaces.

(6) Adding a new capability to be able to do attribute/filters on URI as well as on literals.

(7) Fixing a problem we had when a record had multiple polygons or polylines so that all of them get returned.