(1) Addition of the "attributes_boolean_operator" parameter to the /search/ endpoint. This parameter will tell the endpoint what boolean operator to use ("or" or "and") when doing attribute/value filtering. If "or" is used, this means that if the user filter with 3 attributes, then the returned records will be described using one of these three. If "and" is used, this means that if the user filter with 3 attributes, then the returned records will be described using all the three. This parameter affects all the attribute/value filters. Right now, there is no way to determine the boolean operators between each attribute.
(2) Enabling inference when doing attribute/value filtering when the attribute/value to filter is the core attribute "type". So, if "type" is used as an attribute/value filter, and that inference is "on", then this value will be filtered by inference.

(3) Lat/Long coordinates are now returned even if the endpoint is not geo-enabled. This is important, otherwise if someone define such a lat/long for a record, on a non-geoenabled endpoint, he won't get the lat/long value of that record, even if he was expecting it. This is now fixed.