Addition of a new "target_webservice" parameter.
It determines on what web service URI(s) we should focus on for the listing of the access records. This parameter is used to improve the performance of the web service endpoint depending on the use case. If there are numerous datasets with a numerous number of access permissions defined for each of them, properly using this parameter can have a dramatic impact on the performances. This parameter should be used if the param1 = "access_dataset" or param1 = "access_user" This parameter can have any of these values:

(a) "all" (default): all the web service endpoints URIs for each access records will be taken into account and returned to the user (may be more time consuming).

(b) "none": no web service URI, for any access record, will be returned.

(c) "web-service-uri": any target web service URI that is defined for the list of access records.