I made a mistake when commiting the previous versions of this file.
I did change the "version" attribute of the <schema> element to 1.1 and 1.2.

However, this is not the version attribute of the schema itself, but the version of the parser engine to use to read and process it. Normally, the impact should be minimal, but the problem is that the code has been developed for the schema v1.0 only. The big differences is that in v1.0 of Solr, the multiValued was "true" by default, and in subsequent version is "false". So, this was causing response parsing issues with some endpoint because it was expecting reading a <arr> element for everything (including uri, uid fields etc), but these were not in an <arr> element anymore.