(1) Creation of a new "hierarchy" mode for the getSubClassesDescription() API call.
(2) Creation of a new function: _getClassHierarchyDescription()


NOTE: These new additions are used to create a special purpose web service API call where you want to build a tree hierarchy structure (normally in a UI) that wants to display the structure of the ontology. Previously, it was simply possible with the full description of the class, which may be  quite slow depending on the structure of your ontology.

Check the documentation page for this new API call: http://techwiki.openstructs.org/index.php/Ontology_Read#getSubClasses