Completely changed the way the endpoint was importing/creating ontologies.
The old method was only good if the ontology to import was relatively small (where the size of the OWL file doesn't exceed a few MB). Now, the ontology has two internal modes:

(1) if the ontology file is bigger than 6MB
(2) if the ontology file is smaller than 6MB

In the case (1), the ontology is firstly imported in Virtuoso, and then the structWSF instance get populated.

In the case (2), the same old method is used.

This enhancement has been produced after experimenting some more with bigger ontologies (ones that have 30 000 to 60 0000 resources defined in them).