Instead of using the OWLAPI getOWLClass() and getOWLNamedIndividual(), we are using the getEntitiesInSignature(). That way, we make sure that the entity (the class or the named individual) is in the signature of the ontology.
We used that new API call because we were experiencing issues with the two other calls when the class or the named individual were deleted. These two calls were still returning a reference for the removed entities, when in fact there were none. I am not sure if it is a bug in OWLAPI or simply a strange behavior. Using getEntitiesInSignature() does provide the proper behavior.