Creation of the two new settings:
(1) virtuoso_main_version
(2) enable_lrl

Where:

(1) defines the main version of the Virtuoso server used by this structWSF instance (4, 5 or 6). This is used to know how to use the ODBC API of PHP since its usage differs if we uses version 5 or 6.

(2) enables the Long Read Len feature of Virtuoso if Virtuoso got patched with it. Otherwise, if the length of a field is longer than odbclrl, then it it get trunkated. If this setting is enabled but that Virtuoso is not patched, than this will create inifite loops from time to time.