Major refactory step. All these changes are related to the new internal Resultset structure. This new structure is now used across all the web service endpoints. Now we can say that the data structure used across all the web service endpoints is consistent. The big advantage of this refactoring is that we can *easily* create new core serializations that would be instantly supported by all endpoints.
To create a new serialization, someone would simply have to:

(1) Create a new serialization method in Resultset.php
(2) Add the new mime to Conneg.php
(3) Add the new mime to the accepted mimes of each web service
(4) Handle the new mime in serializations()

The new Resultset and Subject object, used to create the internal array structure, will help developers to create resultsets, and so, to create new web service endpoints.

In the process, we get rid of up to 50% of the code of some of the endpoints. The footprint of structWSF is now much smaller

Also, we optimized the code quite a bit. We minimized the usage of the XML parser (the ParserXML object). We also removed calls to functions that are not needed anymore, with this new refactoring which optimize some of the endpoints.

Overall, this major refactoring activity can be seen as a way to reduce the amount of code, optimize the processes, generalize and make consistent the internal structure used accross all endpoints.