When updating a named individual in the ontology (OWLAPI), we have to make sure to drop all the statements where the value has multiple lines such as:
DataPropertyAssertion(skos:note <http://purl.org/ontology/peg/framework#Energy> "Energy
more text here...
more text here...
"^^xsd:string)

With the previous set of regex rules, this was not matched, and so was re-included after the deletion. The end result is that some triple where supposed to get deleted in the update process, and they were not.