Added a new versioning mechanism to the framework. In this versioning system, two different things got its own version:
  (1) the Web Service endpoints API
  (2) the Sources Interfaces

The goal of the versioning system is to notice the user of the endpoint if
the behavior of the endpoint changed since the client application was developed
for interacting with it.

The version will increase when the behavior of the source interface changes. The
behavior may changes if:

  (1) A bug get fixed (and at least a behavior changes)
  (2) The code get refactored (and at least a behavior changes) [impacts]
  (3) A new parameter/feature is added to the endpoint (but at least another behavior changed) [impacts]
  (4) An existing parameter/feature got changed [impacts]
  (5) An existing parameter/feature got removed [impacts]

In this kind of versioning system, versions are not backward compatible. Because
the version only changes when core behaviors of existing features changes, the
versions can't be backward compatible. It is used to track if behaviors changed,
and to notice users if they did since they last implemented the API.

Also, for the custom source interfaces, if the API of an web service endpoint changed, but
that the source interface hasn't been upgrade to communicate with it, then an error will
be reported as well.