Handling of the minimum cardinality of a property when we index new records we are creating/indexing in structWSF.
If one of the property of a record to create is defined with a sco:maxCardinality = 1 or a sco:cardinality = 1, then we
index it into a *_single_valued field into the Solr index. That way, people are able to specify, in its ontologies
that a property should have a cardinality of 1, which means that they will be sortable by the Search endpoint. This
is how we have to proceed since Solr only support results sorting based on single-valued fields.