Added the "extended_filters" feature/parameter:
Extended filters are used to define more complex search filtered searches. This
parameter uses a more complex syntax which enable the grouping of filter criterias
and the usage of the AND, OR and NOT boolean operators. The grouping is done with
the parenthesis. Each filter is composed of a url-encoded attribute URI to use
as filters, followed by a colomn and the value to filter with. The full lucene
syntax can be used to define the value to filter. If all values are required, the
"*" (start) operator should be used as the value. If the value of an attribute
needs to be considered a URI, then the "[uri]" syntax should be added at the end
of the attribute filter like:
"http%3A%2F%2Fpurl.org%2Fontology%2Ffoo%23friend[uri]:http://bar.com/my-friend-uri".
That way, the value of that attribute filter will be handled as a URI. There are
a series of core attributes that can be used without specifying their full URI:
dataset, type, inferred_type, prefLabel, altLabel, lat, long, description, polygonCoordinates,
polylineCoordinates and located in. The extended filters are not a replacement to
the attributes, types and datasets filtering parameters, they are an extension of it.
Subsequent filtering criterias can be defined in the extended filtering parameter.
The resolution logic by the Search endpoint is:
attributes AND datasets AND types AND extended-filters.
An example of such an extended query is:
(http%3A%2F%2Fpurl.org%2Fontology%2Firon%23prefLabel:cancer AND NOT (breast OR ovarian))
AND (http%3A%2F%2Fpurl.org%2Fontology%2Fnhccn%23useGroupSignificant[uri]:
(http\://purl.org/ontology/doha#liver_cancer OR
http\://purl.org/ontology/doha#cancers_by_histologic_type)) AND
dataset:"file://localhost/data/ontologies/files/doha.owl"