New mechanism to take into account thius scenario:
 (1) An entity get created
 (2) That entity get updated
 (3) That entity get deleted
 (4) A new entity with the same URI get created

Before this path, a 313 error was returned because a revision was existing for this URI but it was not published.

What this update does is to take this usecase into account. Now, the Ontology Update endpoint will catch this usecase, and will re-publish the latest revision, and update it with the input document that as been provided to the Ontology: Update query. This means that no revisions are lost, and that the class will get re-created accordingly.