Fix an issue with the way reasoners where loaded by the OWLOntology() class.
The only way to define the reasoner to use was when the OWLOntology class was created. However, we were specifying it afterward (which had no effect).

This logical code issue got fixed by added a new parameter to the constructor. Now, the  reasoner to use specified in network.ini will be used, and will be used until the ontology get reloaded into the OWLAPI instance.