Changed the behavior of the CRUD: Create mode "searchindex". There was a logic error regarding how this mode was implemented considering the new revisoning feature of the endpoint.
The core issue was that the search index could have been modified without being revisioned. Another issue was that that the solr index could have been updated with unpublished records.

The new behavior of this mode is that only published records can be used in that mode. Also, when a user specify that "searchindex" mode, then the payload of the query is just used to know which record they want to create (update) in the search index. Then, the CRUD: Create endpoint will use the CRUD: Read web service endpoint to get the triples that are currently indexed in the triple store, and will use them to update the search index.

With this new modification, only records that have previously been indexed in the triple store will be indexed into the search index (Solr).