Added a threshold condition to the updateFieldIndex() function
The threshold of maximum 250 fields to check if the field is being used in the index. About that threshold, the system administrator will have to run the Solr Optimize command to optimize the index and remove the un-used fields and then re-run the updateFieldIndex() function. When this threshold is reach, the system assumes that all the fields returned by Luke are being used.

This was causing unacceptable delays with CRUD: Create and CRUD: Update when records were saved that were creating a new field that was not yet in the field index.