New OSF 3.2
This is a major change in OSF. It stopped using ODBC to communicate with the triplestore and now directly communicate with the triple store using the SPARQL endpoint (via HTTP). This is the first path toward having the flexibility to use OSF with any triple stores since this is the most common way to communicate with such a triple store.