Add a new setting to OSF:
    [triplestore]
    virtuoso-disable-transaction-log = "false"

This new setting is used when the triple store is Virtuoso and when we want to disable the transaction log to import new datasets/ontologies. This can be required on a server with billion of triples. In some cases we may have to disable the transaction log such that Virtuoso doesn't "hang" on importing new triples.