Fix a security issue with OSF's core. The issue is that non-OSF datasets that may be existing on a Virtuoso instance used by a OSF instance could be accessed without requiring any user authentication. The OSF queries (using the API Key and API ID) was still required, but the second stage authentication of the user URI was not.
This issue is a security issue only for the instance where such datasets were existing.

A new HTTP error has been created: "WS-AUTH-VALIDATION-104" ; this error is returned when a user try to access a dataset that is not defined in OSF.

This error will be returned by OSF is a dataset URI that is not defined in OSF is used in a OSF query. Even if the dataset is not existing or is empty, this error message will be returned.