Fix an issue with long read length via the ODBC channel.
For some reason, it is possible that odbc_field_len() returns the wrong length (which is lower than the actual length). Because we can't rely on it, we have to find a way to figure out if we are processing a long column. The current way to do this is by checking if the returned value is equal to the LRL limit. If it is, then we recall odbc_result to get the remaining of the value until it returns FALSE.